Move test files for parent directories
