+++ src/moveZeros/__test__/moveZeros.spec.js
@@ -1,4 +1,4 @@
-import {moveZeros} from "../moveZeros/moveZeros.js";
+import {moveZeros} from "../moveZeros.js";
 
 describe('moveZeros', () => {
     it('should move zeros to the end of array', () => {
